+++ packages/client/ui/react-ui/src/components/embed/v3/CrossmintEmbeddedCheckoutV3.tsx
@@ -1,6 +1,6 @@
 import { EmbeddedCheckoutV3IFrame } from "./EmbeddedCheckoutV3IFrame";
-import type { CrossmintEmbeddedCheckoutV3Props } from "@crossmint/client-sdk-base";
+import type { CrossmintEmbeddedCheckoutV3WebProps } from "@crossmint/client-sdk-base";
 
-export function CrossmintEmbeddedCheckout(props: CrossmintEmbeddedCheckoutV3Props) {
+export function CrossmintEmbeddedCheckout(props: CrossmintEmbeddedCheckoutV3WebProps) {
     return <EmbeddedCheckoutV3IFrame {...props} />;
 }