+++ packages/client/window/vitest.config.ts
@@ -1,10 +1,12 @@
+import { resolve } from "path";
 import { defineConfig } from "vitest/config";
 
 export default defineConfig({
     test: {
-        environment: "node",
+        environment: "jsdom",
         include: ["**/*.test.{ts,tsx}"],
         exclude: ["node_modules"],
         globals: true,
+        alias: [{ find: "@", replacement: resolve(__dirname, "./src") }],
     },
 });