+++ packages/server/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @crossmint/server-sdk
 
+## 1.2.80
+
+### Patch Changes
+
+- @crossmint/common-sdk-auth@1.1.16
+
 ## 1.2.79
 
 ### Patch Changes